fix round for usdt

+++ src/bot_app/transactions_usdt.py
@@ -1,6 +1,7 @@
 import logging
 import requests
 
+from decimal import Decimal
 from tronpy import Tron
 from tronpy.keys import PrivateKey
 from tronpy.providers import HTTPProvider
@@ -38,7 +39,7 @@ def create_transaction(dest_address, translation):
     }]
     
     address = pk.public_key.to_base58check_address()
-    amount_in_wei = int(translation * 10 ** 6)  # Convert to USDT's decimal precision (6 decimals)
+    amount_in_wei = Decimal(translation * 10 ** 6)  # Convert to USDT's decimal precision (6 decimals)
     token_contract = tron.get_contract(contract_address)
     token_contract.abi = abi  
     tx = token_contract.functions.transfer(dest_address, amount_in_wei).with_owner(address).fee_limit(50_000_000).build().sign(pk)
